+++ src/pas/plugins/ldap/plonecontrolpanel/__init__.py
@@ -1,12 +1,21 @@
 # -*- coding: utf-8 -*-
-from Products.CMFQuickInstallerTool.interfaces import INonInstallable
+# Products.CMFQuickInstallerTool.interfaces.INonInstallable
+from Products.CMFPlone.interfaces.installable import INonInstallable
 from zope.interface import implementer
 
 
 @implementer(INonInstallable)
 class HiddenProfiles(object):
-    """This hides zope2 profile from the quick installer tool.
+    """This hides zope2 profile from the quick installer tool and plone cpanel
     """
 
+    _hidden = [
+        u'pas.plugins.ldap:default',
+        u'pas.plugins.ldap.plonecontrolpanel:base',
+    ]
+
     def getNonInstallableProducts(self):
-        return ['pas.plugins.ldap']
+        return self._hidden
+
+    def getNonInstallableProfiles(self):
+        return self._hidden